# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Pharmacy Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: List sales
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: userId
          in: query
          schema:
            type: integer
        - name: paymentType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Create a sale (POS)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleInput"
      responses:
        "201":
          description: Created sale with receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      summary: Get a sale with items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"

  # Inventory / Stock Movements
...
components:
  schemas:
...
    SaleItemInput:
      type: object
      required: [productId, quantity, unitPrice]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        unitPrice:
          type: number
    SaleItem:
      type: object
      required: [id, saleId, productId, quantity, unitPrice, subtotal]
      properties:
        id:
          type: integer
        saleId:
          type: integer
        productId:
          type: integer
        productName:
          type: ["string", "null"]
        quantity:
          type: integer
        unitPrice:
          type: number
        subtotal:
          type: number
    Sale:
      type: object
      required: [id, totalAmount, discount, amountPaid, change, paymentType, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        cashierName:
          type: ["string", "null"]
        totalAmount:
          type: number
        discount:
          type: number
        amountPaid:
          type: number
        change:
          type: number
        paymentType:
          type: string
          enum: [cash, wallet, bank]
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        createdAt:
          type: string
    SaleInput:
      type: object
      required: [items, totalAmount, discount, amountPaid, change, paymentType]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItemInput"
        totalAmount:
          type: number
        discount:
          type: number
        amountPaid:
          type: number
        change:
          type: number
        paymentType:
          type: string
          enum: [cash, wallet, bank]

    # Inventory
...