# lib/api-zod/src/generated/types/saleInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pharmacy Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { SaleInputPaymentType } from './saleInputPaymentType';
import type { SaleItemInput } from './saleItemInput';

export interface SaleInput {
  items: SaleItemInput[];
  totalAmount: number;
  discount: number;
  amountPaid: number;
  change: number;
  paymentType: SaleInputPaymentType;
}